Add localization
This commit adds translations of UI parts of app in es_MX and
pt_BR. The home page text has not been translated yet.

The app will always initialize the default currency to devices first
preferred locale. No matter what match the localization framework
determines from the lists of supportedLocales and device locales. This
is implemented with a Future in the MaterialApp constructor, in
localeListResolutionCallback.